Comparing to the original repo, I called the root6 version a different name and reset the file I was working on to be the same as the original repo. Thus, additions to the original repo are now modifications to the .md files only, plus addition of new files.
